Refactoring: corretti riferimenti
